fix(ci): Windows fix

+++ engine/test/engine_tests.cpp
@@ -50,7 +50,7 @@ std::string exec_command(const char *cmd)
     {
         throw std::runtime_error("popen() failed!");
     }
-    while (fgets(buffer.data(), buffer.size(), pipe.get()) != nullptr)
+    while (fgets(buffer.data(), static_cast<int>(buffer.size()), pipe.get()) != nullptr)
     {
         result += buffer.data();
     }
